requirements.txt
llama-index
fastapi
Flask
python-dotenv
pinecone
openai
llama-index-llms-ollama
together
